+++ src/TaskRunner.ts
@@ -1,3 +1,5 @@
+export type OnTaskRunnerMessage = (message: string) => void;
+
 /**
  * The TaskRunner interface is used by the TaskRunnerTerminal to decouple the
  * execution of a task from the VS Code pseudoterminal that runs it.
@@ -8,5 +10,5 @@ export interface TaskRunner {
    *
    * @returns An error message string if the task failed.
    */
-  run(): Promise<void>;
+  run(onMessage: OnTaskRunnerMessage): Promise<void>;
 }